Move server into separate folder
