+++ server/handler/items.go
@@ -6,8 +6,8 @@ import (
     "strconv"
     "github.com/go-chi/chi"
     "github.com/go-chi/render"
-    "github.com/SpennyNDaJets/routine-tracker/db"
-    "github.com/SpennyNDaJets/routine-tracker/models"
+    "github.com/SpennyNDaJets/routine-tracker/server/db"
+    "github.com/SpennyNDaJets/routine-tracker/server/models"
 )
 
 var itemIDKey = "itemID"
